This finishes version 0.1. Initial testing shows both 'pre' and 'post' modes functioning as expected. Further testing is needed though.
This also adds the man 8 page as an uncompressed file. I've added the
undocumented features 'quiet' (-q, --quiet), not quiet (-Q, --not-quiet),
debug (-d, --debug) and disable debug (-D, --no-debug). The default is
no debug and to not be quiet. Debug should be self-explanitory. Quiet
controls the printing of informative message. Critical messages/errors
always print, regardless of 'quiet'. With quiet enabled, there should be
no output from the program.

Signed-off-by: Digimer <digital.mermaid@gmail.com>